tf/requirements.txt
numpy
protobuf
pyyaml
rich
tensorflow-addons
tensorflow[and-cuda]==2.14